+++ Rubedo/Graphics/Particles/Origins/CircleOrigin.cs
@@ -10,9 +10,9 @@ public class CircleOrigin : Origin
     private readonly Interval dist;
     private readonly Interval angle = new Interval(-Math.PI, Math.PI);
     private readonly bool _edge;
-    private readonly int _radius;
+    private readonly float _radius;
 
-    public CircleOrigin(int radius, bool edge = false)
+    public CircleOrigin(float radius, bool edge = false)
     {
         dist = new Interval(0, radius);
         _edge = edge;
@@ -31,7 +31,7 @@ public class CircleOrigin : Origin
         }
         else
         {
-            Vector2 p = new Vector2((int)dist.GetValue(), 0);
+            Vector2 p = new Vector2((float)dist.GetValue(), 0);
             return new OriginData(Vector2.Transform(p, rotation));
         }
     }
